requirements.txt
annotated-types==0.7.0
anyio==4.7.0
click==8.1.7
dependency-injector==4.44.0
fastapi==0.115.6
h11==0.14.0
httptools==0.6.4
idna==3.10
pip==24.3.1
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
pymysql==1.0.3
pipreqs==0.4.10
cryptography==44.0.0
black==24.10.0
isort==5.13.2
pytest==8.3.4
coverage==7.6.9
pytest-cov==6.0.0